requirements.txt
fastapi[standard]
supabase
python-dotenv
google-generativeai